refactor: integrate refreshAllMarks into useSidepanelData for better lifecycle safety (Issue #46)
